Improved perf of collisions

+++ api/src/main/java/com/javaphysicsengine/api/body/PBoundingBox.java
@@ -27,7 +27,7 @@ public class PBoundingBox {
      * @param minY the min Y value
      * @param maxY the max Y value
      */
-    public PBoundingBox(int minX, int maxX, int minY, int maxY) {
+    public PBoundingBox(double minX, double maxX, double minY, double maxY) {
         this.minX = minX;
         this.maxX = maxX;
         this.minY = minY;
